refactor: 파일 업로드 위치 변경
